br/pkg/lightning/log: migrate test-infra to testify (#31289)
close pingcap/tidb#28238
